refactor(relay): merge dto + sputnik into parse/mod.rs
Fold the small dto.rs (request/response types) and sputnik.rs (typed proposal
`kind` view) into parse/mod.rs alongside the parser — they are all "the relay
request and how we read it". wallet_contract stays the one submodule.

dto/sputnik had no external callers outside relay; updated the in-crate import
paths (relay::dto::* / relay::sputnik::* → relay::parse::*). No behavior change.
