--- nt-be/src/handlers/relay/parse/proposal.rs
@@ -1,393 +1,0 @@
-//! The Sputnik DAO proposal operations the relay is willing to sponsor, and the
-//! parser that recovers them from a relayed delegate action.
-//!
-//! Two wire shapes carry the same intent:
-//!
-//! 1. **Direct NEP-366 meta-transaction** — the delegate action's inner
-//!    `FunctionCall` actions call `add_proposal`/`act_proposal` on the
-//!    `*.sputnik-dao.near` treasury directly.
-//! 2. **`w_execute_signed` on a WalletContract** — the same calls live inside the
-//!    wallet request; see [`super::wallet_contract`].
-//!
-//! A single relay is homogeneous: all `add_proposal` or all `act_proposal`. Mixing
-//! the two is rejected, and anything that is not an `add_proposal`/`act_proposal`
-//! targeting the treasury is rejected too. (NEP-141 storage registrations are
-//! derived and paid by the backend separately; they are never relayed as actions.)
-
-use std::ops::Deref;
-
-use near_api::{
-    AccountId, NearToken,
-    types::{Action, transaction::delegate_action::NonDelegateAction},
-};
-use serde::Deserialize;
-use serde_json::Value;
-
-use crate::handlers::relay::confidential::extract_v1_signer_hash_from_kind;
-
-/// The sponsored operation of a single relay. Homogeneous by construction.
-#[derive(Debug, Clone, PartialEq)]
-pub enum RelayOperation {
-    /// One or more `add_proposal` calls.
-    AddProposals(Vec<ProposalInput>),
-    /// One or more `act_proposal` (vote) calls.
-    Votes(Vec<ActProposal>),
-}
-
-/// Which wire shape the relay arrived in.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum RelayShape {
-    /// NEP-366 meta-transaction: the inner actions call the treasury directly.
-    MetaTransaction,
-    /// WalletContract: the actions call `w_execute_signed`, wrapping the proposal
-    /// calls inside the wallet request.
-    WalletContract,
-}
-
-/// The result of parsing a relayed delegate action.
-#[derive(Debug, Clone, PartialEq)]
-pub struct ParsedRelay {
-    pub shape: RelayShape,
-    pub operation: RelayOperation,
-    /// Total NEAR attached across the sponsored calls (proposal bonds).
-    pub attached_deposit: NearToken,
-}
-
-/// The `proposal` argument of `add_proposal`. `kind` is kept as raw JSON because the
-/// relay does not authorize on it — DAO permissions are enforced on-chain and via
-/// the permission checks in [`super::access`].
-#[derive(Debug, Clone, PartialEq, Deserialize)]
-pub struct ProposalInput {
-    #[serde(default)]
-    pub description: String,
-    pub kind: Value,
-}
-
-/// The arguments of `act_proposal`.
-#[derive(Debug, Clone, PartialEq)]
-pub struct ActProposal {
-    pub id: u64,
-    pub action: String,
-    /// Client-provided proposal `kind` echoed in the `act_proposal` args. Used only
-    /// to surface confidential payload hashes, never for authorization.
-    pub kind: Option<Value>,
-}
-
-impl RelayOperation {
-    /// Whether this relay adds proposals. Only `add_proposal` grows the DAO
-    /// contract's storage, so only then does the relayer top up its balance.
-    pub fn is_add_proposals(&self) -> bool {
-        matches!(self, RelayOperation::AddProposals(_))
-    }
-
-    /// Proposal ids being approved (`VoteApprove`), in order. Empty for add relays.
-    pub fn vote_approve_ids(&self) -> Vec<u64> {
-        match self {
-            RelayOperation::Votes(votes) => votes
-                .iter()
-                .filter(|vote| vote.action == "VoteApprove")
-                .map(|vote| vote.id)
-                .collect(),
-            RelayOperation::AddProposals(_) => Vec::new(),
-        }
-    }
-
-    /// Confidential `v1.signer` payload hashes referenced by the votes, in order. A
-    /// single relay can carry several (e.g. multiple confidential votes). Empty for
-    /// add relays.
-    pub fn confidential_payload_hashes(&self) -> Vec<String> {
-        match self {
-            RelayOperation::Votes(votes) => votes
-                .iter()
-                .filter_map(|vote| {
-                    vote.kind
-                        .as_ref()
-                        .and_then(extract_v1_signer_hash_from_kind)
-                })
-                .collect(),
-            RelayOperation::AddProposals(_) => Vec::new(),
-        }
-    }
-}
-
-/// Parse a relayed delegate action into the operation to sponsor, rejecting anything
-/// that is not a homogeneous set of `add_proposal`/`act_proposal` calls targeting
-/// `treasury_id`.
-///
-/// `action_receiver_id` is the delegate action's receiver — the contract the inner
-/// actions call in the direct shape. For `w_execute_signed` the receiver is read
-/// from each inner promise instead, so this argument is ignored there.
-pub fn parse_sponsored_proposals(
-    treasury_id: &AccountId,
-    action_receiver_id: &AccountId,
-    actions: &[NonDelegateAction],
-) -> Result<ParsedRelay, String> {
-    let (shape, calls) = if is_wallet_contract_shape(actions) {
-        (
-            RelayShape::WalletContract,
-            super::wallet_contract::collect_calls(actions)?,
-        )
-    } else {
-        (
-            RelayShape::MetaTransaction,
-            collect_direct_calls(action_receiver_id, actions)?,
-        )
-    };
-    let (operation, attached_deposit) = validate_calls(treasury_id, calls)?;
-    Ok(ParsedRelay {
-        shape,
-        operation,
-        attached_deposit,
-    })
-}
-
-/// A relay is WalletContract-shaped when its actions call `w_execute_signed`. We peek
-/// the first action to route; [`super::wallet_contract::collect_calls`] then validates
-/// every action (via `as_w_execute_call`), rejecting a mixed delegate action.
-fn is_wallet_contract_shape(actions: &[NonDelegateAction]) -> bool {
-    matches!(
-        actions.first().map(Deref::deref),
-        Some(Action::FunctionCall(fc)) if fc.method_name == "w_execute_signed"
-    )
-}
-
-/// A single contract call extracted from either wire shape, normalized so the two
-/// flows validate identically.
-#[derive(Debug, Clone, PartialEq)]
-pub(crate) struct DaoCall {
-    pub(crate) receiver_id: AccountId,
-    pub(crate) method_name: String,
-    /// Raw JSON argument bytes.
-    pub(crate) args: Vec<u8>,
-    pub(crate) deposit: NearToken,
-}
-
-/// Direct meta-transaction: every inner action must be a `FunctionCall`. They all
-/// share the delegate action's single `receiver_id`, which [`validate_calls`] then
-/// pins to the treasury.
-fn collect_direct_calls(
-    receiver_id: &AccountId,
-    actions: &[NonDelegateAction],
-) -> Result<Vec<DaoCall>, String> {
-    actions
-        .iter()
-        .map(|action| match action.deref() {
-            Action::FunctionCall(fc) => Ok(DaoCall {
-                receiver_id: receiver_id.clone(),
-                method_name: fc.method_name.clone(),
-                args: fc.args.clone(),
-                deposit: fc.deposit,
-            }),
-            _ => Err("Delegate action contains a non-FunctionCall action".to_owned()),
-        })
-        .collect()
-}
-
-/// Validate the normalized calls: each must target the treasury and be an
-/// `add_proposal`/`act_proposal`, and the relay must be homogeneous (all of one
-/// kind). Returns the sponsored operation plus the total attached deposit.
-fn validate_calls(
-    treasury_id: &AccountId,
-    calls: Vec<DaoCall>,
-) -> Result<(RelayOperation, NearToken), String> {
-    if calls.is_empty() {
-        return Err("Delegate action contains no sponsorable proposal calls".to_owned());
-    }
-
-    let mut add_proposals = Vec::new();
-    let mut votes = Vec::new();
-    let mut attached_deposit = NearToken::from_near(0);
-    for call in calls {
-        if &call.receiver_id != treasury_id {
-            return Err(format!(
-                "Proposal call targets '{}', but only the treasury '{}' may be sponsored",
-                call.receiver_id, treasury_id
-            ));
-        }
-        attached_deposit = attached_deposit.saturating_add(call.deposit);
-        match call.method_name.as_str() {
-            "add_proposal" => add_proposals.push(parse_add_proposal(&call)?),
-            "act_proposal" => votes.push(parse_act_proposal(&call)?),
-            other => {
-                return Err(format!(
-                    "Unsupported relayed method '{}' (only add_proposal/act_proposal are sponsored)",
-                    other
-                ));
-            }
-        }
-    }
-
-    let operation = match (add_proposals.is_empty(), votes.is_empty()) {
-        (false, true) => RelayOperation::AddProposals(add_proposals),
-        (true, false) => RelayOperation::Votes(votes),
-        // Both non-empty: a mix. (Both empty is impossible — calls is non-empty and
-        // every call is one kind or the other.)
-        _ => {
-            return Err(
-                "A relay must contain only add_proposal or only act_proposal calls, not a mix"
-                    .to_owned(),
-            );
-        }
-    };
-
-    Ok((operation, attached_deposit))
-}
-
-fn parse_add_proposal(call: &DaoCall) -> Result<ProposalInput, String> {
-    let args: AddProposalArgs = serde_json::from_slice(&call.args)
-        .map_err(|e| format!("Invalid add_proposal args: {}", e))?;
-    Ok(args.proposal)
-}
-
-fn parse_act_proposal(call: &DaoCall) -> Result<ActProposal, String> {
-    let args: ActProposalArgs = serde_json::from_slice(&call.args)
-        .map_err(|e| format!("Invalid act_proposal args: {}", e))?;
-    Ok(ActProposal {
-        id: args.id,
-        action: args.action,
-        kind: args.proposal,
-    })
-}
-
-#[derive(Deserialize)]
-struct AddProposalArgs {
-    proposal: ProposalInput,
-}
-
-#[derive(Deserialize)]
-struct ActProposalArgs {
-    id: u64,
-    action: String,
-    #[serde(default)]
-    proposal: Option<Value>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use serde_json::json;
-
-    const TREASURY: &str = "dao.sputnik-dao.near";
-
-    fn acc(s: &str) -> AccountId {
-        s.parse().unwrap()
-    }
-
-    fn b64(value: &Value) -> String {
-        use base64::Engine as _;
-        base64::engine::general_purpose::STANDARD.encode(serde_json::to_vec(value).unwrap())
-    }
-
-    fn add_call(treasury: &str, kind: Value, deposit_yocto: u128) -> DaoCall {
-        DaoCall {
-            receiver_id: acc(treasury),
-            method_name: "add_proposal".to_owned(),
-            args: serde_json::to_vec(&json!({
-                "proposal": { "description": "d", "kind": kind }
-            }))
-            .unwrap(),
-            deposit: NearToken::from_yoctonear(deposit_yocto),
-        }
-    }
-
-    fn act_call(treasury: &str, id: u64, action: &str, kind: Option<Value>) -> DaoCall {
-        let mut args = json!({ "id": id, "action": action });
-        if let Some(kind) = kind {
-            args["proposal"] = kind;
-        }
-        DaoCall {
-            receiver_id: acc(treasury),
-            method_name: "act_proposal".to_owned(),
-            args: serde_json::to_vec(&args).unwrap(),
-            deposit: NearToken::from_near(0),
-        }
-    }
-
-    fn transfer_kind() -> Value {
-        json!({ "Transfer": { "token_id": "", "receiver_id": "bob.near", "amount": "1" } })
-    }
-
-    #[test]
-    fn validates_add_proposals_and_sums_deposit() {
-        let (operation, attached_deposit) = validate_calls(
-            &acc(TREASURY),
-            vec![add_call(TREASURY, transfer_kind(), 100)],
-        )
-        .unwrap();
-        match operation {
-            RelayOperation::AddProposals(ref inputs) => assert_eq!(inputs.len(), 1),
-            other => panic!("expected AddProposals, got {other:?}"),
-        }
-        assert_eq!(attached_deposit, NearToken::from_yoctonear(100));
-    }
-
-    #[test]
-    fn extracts_vote_approve_ids() {
-        let calls = vec![
-            act_call(TREASURY, 7, "VoteApprove", None),
-            act_call(TREASURY, 8, "VoteReject", None),
-            act_call(TREASURY, 9, "VoteApprove", None),
-        ];
-        let (operation, _) = validate_calls(&acc(TREASURY), calls).unwrap();
-        assert_eq!(operation.vote_approve_ids(), vec![7, 9]);
-    }
-
-    #[test]
-    fn extracts_multiple_confidential_hashes() {
-        let v1_kind = |hash: &str| {
-            json!({
-                "FunctionCall": {
-                    "receiver_id": "v1.signer",
-                    "actions": [{
-                        "method_name": "sign",
-                        "args": b64(&json!({ "request": { "payload_v2": { "Eddsa": hash } } })),
-                    }]
-                }
-            })
-        };
-        let calls = vec![
-            act_call(TREASURY, 1, "VoteApprove", Some(v1_kind("aaaa"))),
-            act_call(TREASURY, 2, "VoteApprove", Some(v1_kind("bbbb"))),
-        ];
-        let (operation, _) = validate_calls(&acc(TREASURY), calls).unwrap();
-        assert_eq!(
-            operation.confidential_payload_hashes(),
-            vec!["aaaa".to_owned(), "bbbb".to_owned()]
-        );
-    }
-
-    #[test]
-    fn rejects_mixed_add_and_act() {
-        let calls = vec![
-            add_call(TREASURY, transfer_kind(), 0),
-            act_call(TREASURY, 1, "VoteApprove", None),
-        ];
-        let err = validate_calls(&acc(TREASURY), calls).unwrap_err();
-        assert!(
-            err.contains("only add_proposal or only act_proposal"),
-            "{err}"
-        );
-    }
-
-    #[test]
-    fn rejects_non_proposal_method() {
-        let call = DaoCall {
-            receiver_id: acc(TREASURY),
-            method_name: "storage_deposit".to_owned(),
-            args: b"{}".to_vec(),
-            deposit: NearToken::from_near(0),
-        };
-        let err = validate_calls(&acc(TREASURY), vec![call]).unwrap_err();
-        assert!(err.contains("Unsupported relayed method"), "{err}");
-    }
-
-    #[test]
-    fn rejects_call_to_other_receiver() {
-        let err = validate_calls(
-            &acc(TREASURY),
-            vec![add_call("evil.near", transfer_kind(), 0)],
-        )
-        .unwrap_err();
-        assert!(err.contains("only the treasury"), "{err}");
-    }
-}
